pyproject.toml
[tool.poetry]
name = "eb2"
version = "0.1.0"
description = ""
authors = ["Luiz A F Gomes"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
profanity-check = "^1.0.3"
colorama = "^0.4.3"
django = "^3.0.5"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
